Fix dtype mismatch by ensuring float32 throughout diffusion process
This commit fixes the RuntimeError: mat1 and mat2 must have the same dtype
error by ensuring float32 is used consistently throughout the diffusion
sampling process.

Changes:
- Modified calc_diffusion_step_embedding() to use torch.arange() with
  dtype=torch.float32 instead of default float64
- Changed diffusion_steps creation to use torch.full() with explicit
  float32 dtype instead of multiplication which could result in type
  promotion to float64
- Applied changes to both src/sssd and sssd_standalone directories

This ensures that all tensors used in the model forward pass are float32,
matching the model's weight precision and preventing dtype mismatches.

+++ src/sssd/utils/util.py
@@ -79,7 +79,7 @@ def calc_diffusion_step_embedding(diffusion_steps, diffusion_step_embed_dim_in):
 
     half_dim = diffusion_step_embed_dim_in // 2
     _embed = np.log(10000) / (half_dim - 1)
-    _embed = torch.exp(torch.arange(half_dim) * -_embed).cuda()
+    _embed = torch.exp(torch.arange(half_dim, dtype=torch.float32) * -_embed).cuda()
     _embed = diffusion_steps * _embed
     diffusion_step_embed = torch.cat((torch.sin(_embed),
                                       torch.cos(_embed)), 1)
